pyproject.toml
[project]
name = "databridge-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databridge-client>=0.2.0",
    "mcp[cli]>=1.2.1",
    "ruff>=0.9.6",
]
